+++ node/clients/Organizations.ts
@@ -1,13 +1,8 @@
-import type { GraphQLResponse, InstanceOptions, IOContext } from '@vtex/api'
+import type { InstanceOptions, IOContext } from '@vtex/api'
 import { AppGraphQLClient } from '@vtex/api'
 
 import { QUERIES } from '../resolvers/Routes/utils'
 import { getTokenToHeader } from './index'
-import type {
-  GetCostCenterType,
-  GetOrganizationsByEmailResponse,
-  GetOrganizationsPaginatedByEmailResponse,
-} from '../typings/custom'
 
 const getPersistedQuery = () => {
   return {
@@ -41,16 +36,6 @@ export class OrganizationsGraphQLClient extends AppGraphQLClient {
     })
   }
 
-  public getCostCenterById = async (costId: string) => {
-    return this.query({
-      extensions: getPersistedQuery(),
-      query: QUERIES.getCostCenterById,
-      variables: {
-        id: costId,
-      },
-    }) as Promise<GraphQLResponse<GetCostCenterType>>
-  }
-
   public getMarketingTags = async (costId: string): Promise<unknown> => {
     return this.query({
       extensions: getPersistedQuery(),
@@ -59,34 +44,6 @@ export class OrganizationsGraphQLClient extends AppGraphQLClient {
         costId,
       },
     })
-  }
-
-  public getOrganizationsByEmail = async (email: string) => {
-    return this.query({
-      extensions: getPersistedQuery(),
-      query: QUERIES.getOrganizationsByEmail,
-      variables: { email },
-    }) as Promise<GetOrganizationsByEmailResponse>
-  }
-
-  public getOrganizationsPaginatedByEmail = async (
-    email: string,
-    page: number,
-    pageSize: number
-  ) => {
-    return this.query({
-      extensions: getPersistedQuery(),
-      query: QUERIES.getOrganizationsPaginatedByEmail,
-      variables: {
-        email,
-        page,
-        pageSize,
-      },
-    }) as Promise<{
-      data: {
-        getOrganizationsPaginatedByEmail: GetOrganizationsPaginatedByEmailResponse
-      }
-    }>
   }
 
   private query = async (param: {